Fix #4690, duplicated property to single-property Record
