+++ src/test/java/tools/jackson/databind/records/DuplicatePropertyDeserializationRecord4690Test.java
@@ -1,0 +1,64 @@
+package tools.jackson.databind.records;
+
+import org.junit.jupiter.api.Test;
+
+import tools.jackson.databind.ObjectMapper;
+import tools.jackson.databind.testutil.DatabindTestUtil;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+// [databind#4690] InvalidDefinitionException "No fallback setter/field defined for creator property"
+// when deserializing JSON with duplicated property to single-property Record
+public class DuplicatePropertyDeserializationRecord4690Test
+        extends DatabindTestUtil
+{
+    record MyRecord(String first) { }
+
+    static class MyClass {
+        private String first;
+
+        MyClass(String first) {}
+
+        public void setFirst(String first) {
+            this.first = first;
+        }
+
+        public String getFirst() {
+            return first;
+        }
+    }
+
+    private final ObjectMapper mapper = newJsonMapper();
+
+    @Test
+    void testDuplicatePropertyDeserialization() throws Exception {
+        final String json = a2q("{'first':'value','first':'value2'}");
+
+        MyRecord result = mapper.readValue(json, MyRecord.class);
+
+        assertNotNull(result);
+        assertEquals("value2", result.first());
+    }
+
+    @Test
+    void testDuplicatePropertyDeserialization2() throws Exception {
+        final String json = a2q("{'first':'value','second':'test1','first':'value2'}");
+
+        MyRecord result = mapper.readValue(json, MyRecord.class);
+
+        assertNotNull(result);
+        assertEquals("value2", result.first());
+    }
+
+    @Test
+    void testDuplicatePropertyClassDeserialization() throws Exception {
+        final String json = a2q("{'first':'value','second':'test1','first':'value2'}");
+
+        MyClass result = mapper.readValue(json, MyClass.class);
+
+        assertNotNull(result);
+        assertEquals("value2", result.getFirst());
+    }
+
+}
